+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Updated to tinted-builder `0.16.0` lib
+
 ## [0.19.0] - 2026-04-30
 
 ### Changed